Added extra checks, including syscall fake hook
To prevent syscall, search for the byte pattern 50 2A 80 D2 01 10 00 D4 in a hex editor (mov x16, 338; svc 0x80) in the case of arm64 and replace it with 1F 20 03 D5 1F 20 03 D5 to nop it. In v1.5.1, this is at .text:10033BDB0). I do not have access to other versions or archs but in 32-bit ARM the byte pattern should be 52 C1 00 E3 80 00 00 EF (mov r12, 338; svc 0x80) and should probably replaced with 00 F0 20 E3 00 F0 20 E3 to nop it. If this isn't the case or it does not work, please create an issue in the repository