Last indefinite cleanup, commenting out syscall fake hook
To prevent syscall, search for the byte pattern 30 04 80 D2 01 10 00 D4 in a hex editor (mov x16, 33; svc 0x80) in the case of arm64 and replace it with 1F 20 03 D5 1F 20 03 D5 to nop it. In my version dumped at 6/25 with the binary (/en.app/en) with a sha256 of 94d63bed366350c82a2ee7a7134315866c65807cbcc09abd3ad04702102e705c, this is at .text:100F385EC). I couldn't even get this app to load complete on my non-jailbroken 4s so I can't confirm for 32-bit. If any of this is not the case or it does not work, please create an issue in the repository